Update BAM_NN.py
The object of development in this MD is the Development of a neural network of bidirectional associative memory for digit recognition. The purpose of this work is to write and debug software code that allows receiving data and recognizing digits. The results can be used to recognize numbers in various areas of educational and law enforcement activities.